// apps/frontend/src/api/store/types.ts
export interface StoreUpdateInformationForm {
  name?: string;
  whatsapp?: string;
  telephone?: string;
  category?: string;
  currency?: string;
  logo_img?: string;
  banner_img?: string;
  description?: string;
  team_img?: string;
  team_description?: string;
}

export interface StoreUpdateSocialForm {
  facebook: string;
  instagram: string;
  pinterest: string;
  twitter: string;
  tiktok: string;
  youtube: string;
}